+++ CHANGELOG.md
@@ -63,3 +63,9 @@
 ### Adicionando
 
 - Conexão de botão de editar avaliação com modal de edição
+
+## US7.4 - PR 16
+
+### Adicionando 
+
+- Tela de exibir estabelecimento